+++ packages/cli/src/firebase.ts
@@ -1,6 +1,6 @@
 import { initializeApp } from 'firebase/app';
 import { initializeAuth } from 'firebase/auth';
-import { getFirestore } from 'firebase/firestore/lite';
+import { getFirestore } from 'firebase/firestore';
 
 export const firebaseConfig = {
   apiKey: "AIzaSyCz2PtDtfSybjPB5y0eGw4-gh3Aic55NFk",