+++ packages/framework/src/hooks/useThemeContext.tsx
@@ -1,14 +1,14 @@
-import { useAtom } from "jotai";
-import { createContext } from "react";
+import { useAtom, useAtomValue } from "jotai";
+import { keys } from "lodash-es";
+import { createContext, useCallback } from "react";
 import { type EnsembleThemeModel } from "../shared";
-import { defaultThemeDefinition, themeAtom } from "../state";
+import { appAtom, defaultThemeDefinition, themeAtom } from "../state";
 
 export type CustomTheme = { [key: string]: unknown } & {
   theme: EnsembleThemeModel;
   themeName?: string;
   setTheme?: (name: string) => void;
 };
-
 export const CustomThemeContext = createContext<CustomTheme>({
   theme: defaultThemeDefinition,
 });
@@ -18,11 +18,21 @@ export const useThemeScope = (): {
   themeName: string;
   setTheme: (name: string) => void;
 } => {
-  const [theme, setTheme] = useAtom(themeAtom);
+  const [theme, updateTheme] = useAtom(themeAtom);
 
-  return {
-    theme,
-    themeName: theme.name,
-    setTheme,
-  };
+  const appContext = useAtomValue(appAtom);
+
+  const setTheme = useCallback(
+    (name: string) => {
+      if (keys(appContext.application?.themes).includes(name)) {
+        const selectedTheme = appContext.application?.themes[name];
+        if (selectedTheme) {
+          updateTheme(selectedTheme);
+        }
+      }
+    },
+    [updateTheme, appContext.application?.themes],
+  );
+
+  return { theme, themeName: theme.name, setTheme };
 };