+++ packages/js-commons/src/browser/index.ts
@@ -1,0 +1,2 @@
+export * from "../core/firebase";
+export * from "../core/service";