+++ packages/framework/CHANGELOG.md
@@ -1,11 +1,10 @@
 # @ensembleui/react-framework
 
-## 0.2.9
+## 0.2.10
 
 ### Patch Changes
 
-- 5094d08: fix: refactor missing themeContext in useCommandCallback
-- ec9d081: added support for cache in API call
+- 6cbe562: Revert hook changes
 
 ## 0.2.8